Create medizinische-fakultat-mannheim-numerisch.csl
A modified version of the "Journal of the American Society of Nephrology"-Style for the purposes of the Mannheim Medical School by the University of Heidelberg (Medizinische Fakultät Mannheim der Universität Heidelberg) in Germany.

KNOWN BUG: Usage of the style causes Zotero to fail updating the references in a document when a web page is present in the references.  CSL editor shows error message: "Citeproc exception: TypeError: Cannot read property 'strings' of undefined". Error not identified yet, maybe someone with more profound skills can help here?